fix view
