+++ Tasks/Test_task_1.txt
@@ -1,0 +1,28 @@
+Самостоятельная работа вер.1
+
+До конца пары показать, иначе -1.
+
+Постановка задачи. 
+Реализовать класс Расстояние (Distance).
+Класс должен уметь работать с расстояниями в метрах, километрах, милях.
+Значения хранятся в целых числах.
+
+Требования.
+1. Не хранить все 3 расстояния. В качестве поля использовать только то, что считается общепризнанной мерой расстояния
+в системе СИ - метры.
+2. Минимальный набор функций:
+- базовые конструкторы: конструктор по умолчанию, конструктор инициализации,
+- конструкторы преобразования типа: из километров, из миль.
+- перегрузки основных арифметических операторов:
+	+(к расстоянию можно прибавлять расстояния, мили, километры, метры),
+	-= (из расстояния можно вычитать расстояния, мили, километры, метры), 
+	* на число.
+- перегрузка оператора присваивания =
+- перегрузка операторов сравнения двух расстояний: ==, !=, >, <, >=, <=
+- функции to_miles(), to_kilometers() для перевода расстояний к указанным типам.
+- вывод на экран (метод или перегрузка оператора <<)
+
+Подсказки.
+1. Для удобства работы можно реализовать структуры/классы miles_t, kilometers_t, meters_t.
+2. Методы можно реализовывать через уже написанные, например, != можно реализовать через вызов ==.
+в реализации += использовать уже реализованный +.
