+++ Tasks/Test_task_2.txt
@@ -1,0 +1,25 @@
+Самостоятельная работа вер.2
+
+До конца пары, иначе -1.
+
+Постановка задачи. 
+Реализовать класс Вес (Weight).
+Класс должен уметь работать с весами в килограммах, граммах, фунтах.
+
+Требования.
+1. Не хранить все 3 веса. В качестве поля использовать только то, что считается общепризнанной мерой массы
+в системе СИ - граммы.
+2. Минимальный набор функций:
+- базовые конструкторы: конструктор по умолчанию, конструктор инициализации;
+- конструкторы преобразования типа: из килограммов, из фунтов;
+- перегрузки основных арифметических операторов:
+	-(к весу можно прибавлять веса, фунты, килограммы, граммы),
+	+= (из расстояния можно вычитать веса, фунты, килограммы, граммы), 
+	* на число;
+- перегрузка оператора присваивания =;
+- перегрузка операторов сравнения двух весов: ==, !=, >, <, >=, <=;
+- функции to_pounds(), to_kilograms() для перевода расстояний к указанным типам;
+- вывод на экран (метод или перегрузка оператора <<).
+
+Подсказки.
+1. Для удобства работы можно реализовать структуры/классы pounds_t, kilograms_t, grams_t.
